requirements.txt
amqp==5.0.6
celery==5.1.2
Django==3.2.8
django-admin-interface==0.17.3
django-celery-results==2.2.0
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-filter==21.1
django-nested-admin==3.3.3
djangorestframework==3.12.4
django-storages==1.12.1
django-widget-tweaks==1.4.8
psycopg2==2.9.1 --no-binary psycopg2
pur==5.4.2
pylint==2.11.1
pylint-django==2.4.4
pylint-fail-under==0.3.0
pylint-plugin-utils==0.6
requests==2.26.0
sentry-sdk==1.4.3
URLObject==2.4.3
wheel==0.37.0
